+++ examples/inverse/plot_lcmv_beamformer_volume.py
@@ -101,8 +101,7 @@ stc.save('lcmv-vol')
 stc.crop(0.0, 0.2)
 
 # Save result in a 4D nifti file
-img = mne.save_stc_as_volume('lcmv_inverse.nii.gz', stc,
-                             forward['src'], mri_resolution=False)
+img = stc.as_volume(forward['src'], mri_resolution=False)
 
 t1_fname = data_path + '/subjects/sample/mri/T1.mgz'