+++ mne/minimum_norm/inverse.py
@@ -947,7 +947,8 @@ def apply_inverse(evoked, inverse_operator, lambda2=1. / 9., method="dSPM",
     subject = _subject_from_inverse(inverse_operator)
 
     stc = _make_stc(sol, vertno, tmin=tmin, tstep=tstep, subject=subject,
-                    vector=(pick_ori == 'vector'), source_nn=source_nn)
+                    vector=(pick_ori == 'vector'), source_nn=source_nn,
+                    src=inverse_operator['src'])
     logger.info('[done]')
 
     return stc
@@ -1086,7 +1087,8 @@ def apply_inverse_raw(raw, inverse_operator, lambda2, method="dSPM",
     tstep = 1.0 / raw.info['sfreq']
     subject = _subject_from_inverse(inverse_operator)
     stc = _make_stc(sol, vertno, tmin=tmin, tstep=tstep, subject=subject,
-                    vector=(pick_ori == 'vector'), source_nn=source_nn)
+                    vector=(pick_ori == 'vector'), source_nn=source_nn,
+                    src=inverse_operator['src'])
     logger.info('[done]')
 
     return stc
@@ -1153,7 +1155,8 @@ def _apply_inverse_epochs_gen(epochs, inverse_operator, lambda2, method='dSPM',
                 sol = np.dot(K, e[sel])
 
         stc = _make_stc(sol, vertno, tmin=tmin, tstep=tstep, subject=subject,
-                        vector=(pick_ori == 'vector'), source_nn=source_nn)
+                        vector=(pick_ori == 'vector'), source_nn=source_nn,
+                        src=inverse_operator['src'])
 
         yield stc
 
@@ -1226,6 +1229,7 @@ def apply_inverse_epochs(epochs, inverse_operator, lambda2, method="dSPM",
     return stcs
 
 
+# XXX what is this???
 '''
 def _xyz2lf(Lf_xyz, normals):
     """Reorient leadfield to one component matching the normal to the cortex
@@ -1722,11 +1726,12 @@ def estimate_snr(evoked, inv, verbose=None):
 
     # Adapted from mne_analyze/regularization.c, compute_regularization
     n_zero = (inv['noise_cov']['eig'] <= 0).sum()
+    n_ch_eff = n_ch - n_zero
     logger.info('Effective nchan = %d - %d = %d'
-                % (n_ch, n_zero, n_ch - n_zero))
+                % (n_ch, n_zero, n_ch_eff))
+    del n_ch
     signal = np.sum(data_white ** 2, axis=0)  # sum of squares across channels
-    noise = n_ch - n_zero
-    snr = signal / noise
+    snr = signal / n_ch_eff
 
     # Adapted from noise_regularization
     lambda2_est = np.empty(n_times)
@@ -1735,13 +1740,13 @@ def estimate_snr(evoked, inv, verbose=None):
 
     # deal with low SNRs
     bad = (snr <= 1)
-    lambda2_est[bad] = 100.
+    lambda2_est[bad] = np.inf
     remaining[bad] = False
 
     # parameters
-    lambda_mult = 0.9
+    lambda_mult = 0.99
     sing2 = (inv['sing'] * inv['sing'])[:, np.newaxis]
-    val = chi2.isf(1e-3, n_ch - 1)
+    val = chi2.isf(1e-3, n_ch_eff)
     for n_iter in range(1000):
         # get_mne_weights (ew=error_weights)
         # (split newaxis creation here for old numpy)